requirements.txt
astroid==2.2.5
atomicwrites==1.3.0
attrs==19.1.0
certifi==2019.6.16
chardet==3.0.4
coverage==4.5.3
coveralls==1.8.1
Django==2.2.2
djangorestframework==3.9.4
djangorestframework-simplejwt==4.3.0
docopt==0.6.2
factory-boy==2.12.0
Faker==1.0.7
idna==2.8
importlib-metadata==0.17
isort==4.3.21
lazy-object-proxy==1.4.1
mccabe==0.6.1
more-itertools==7.0.0
packaging==19.0
pluggy==0.12.0
psycopg2==2.7.7
py==1.8.0
PyJWT==1.7.1
pylint==2.3.1
pyparsing==2.4.0
pytest==4.6.2
python-dateutil==2.8.0
pytz==2019.1
requests==2.22.0
six==1.12.0
sqlparse==0.3.0
text-unidecode==1.2
typed-ast==1.4.0
urllib3==1.25.3
wcwidth==0.1.7
wrapt==1.11.2
zipp==0.5.1
